Sketch breaking rivers up into smaller patches
This is necessary to avoid large chunks of background because the rivers
cover such a a large area that when we pick a chip inside the bounds of
the river, it is often completely background.

This commit is not finished and includes a bunch of print statements
that need to be removed; for debugging, testing and KC labels are
disabled.