Fix bug when DEM is disabled
Previously we were hardcoding the number of input channels to the model.
We need it to adapt depending on whether or not DEM and NIR are enabled.
This required refactoring the create_model function; instead of passing
a million arguments and two new ones, we just pass the config itself and
read from it.